refactor: SM-216 Image 컴포넌트 관련 컴포넌트들 prop 상수화
