+++ src/components/compound/Loader/index.tsx
@@ -1,10 +1,16 @@
 import type { ReactElement } from 'react';
 import { StyledImage } from './styled';
 import imageSrc from '@assets/loader/shaker.png';
+import { IMG_MODE } from '@constants';
 
 const Loader = (): ReactElement => {
   return (
-    <StyledImage height='80px' mode='contain' src={imageSrc} width='80px' />
+    <StyledImage
+      height='80px'
+      mode={IMG_MODE.CONTAIN}
+      src={imageSrc}
+      width='80px'
+    />
   );
 };
