+++ src/pages/ErrorPage/index.tsx
@@ -5,6 +5,7 @@ import { Text, Image } from '@base';
 import { TextButton } from '@compound';
 import { HeaderPageTemplate } from '@domain';
 import notFoundPage from '@assets/notFoundPage.png';
+import { IMG_MODE } from '@constants';
 
 const ErrorPage = (): ReactElement => {
   const navigate = useNavigate();
@@ -16,7 +17,7 @@ const ErrorPage = (): ReactElement => {
   return (
     <HeaderPageTemplate>
       <StyledContainer>
-        <Image mode='contain' src={notFoundPage} />
+        <Image mode={IMG_MODE.CONTAIN} src={notFoundPage} />
         <Text>페이지를 찾을 수 없습니다!</Text>
         <TextButton buttonType='LONG_PINK' onClick={handleClick}>
           Go to Main Page
